Addition of Editor trend sub conditions. Improvement of request/task calculation and notification. Fix TALIB missing operations and descriptions. Divers bug fixes

+++ premiumMarkets/pm-core/src/main/java/com/finance/pms/events/scoring/chartUtils/BarSettings.java
@@ -40,7 +40,7 @@ public class BarSettings {
 	}
 	public void setMaxFill(int maxFill) {
 		try {
-			MainPMScmd.getPrefs().putDouble("chart.maxFill", maxFill);
+			MainPMScmd.getPrefs().putInt("chart.maxFill", maxFill);
 		} catch (Exception e) {
 			LOGGER.error(e, e);
 		}
